requirements.txt
numpy==1.19.2
tqdm==4.50.2
streamlit==0.65.0
streamlit_drawable_canvas==0.5.1
tensorflow==2.3.1
matplotlib==3.3.2
ipython==7.18.1
Pillow==8.0.0
opencv-python==3.4.2
